Stop all the "async" and lazyloading nonsense
According to some testing, lazyloading plugins like
feline, startify, lspconfig and others by using the
"async" function doesn't really improve anything
(startuptime). In fact, even if the log from running nvim with
--startuptime flag says that the its faster, it actually feels slower
since there's a small delay before the statusline and startify (and
sometimes the colorscheme) get triggered. While when those plugins are
not lazyloaded, it basically happens instantly without any delay (or at
least that's how it feels).

More interesting is the fact that benchmarking with hyperfine, shows
that the startuptime increases when using the "async" function,
(contrary to what the --startuptime flag says).

Results of some of my testing:

* When lazy loading with "async".

 hyperfine "nvim +q" --warmup 40
    Benchmark #1: nvim +q
        Time (mean ± σ):      87.3 ms ±   5.3 ms    [User: 71.0 ms, System: 14.0 ms]
        Range (min … max):    79.6 ms … 102.2 ms    45 runs

* When not using "async"

hyperfine "nvim +q"

    Benchmark #1: nvim +q
      Time (mean ± σ):      66.2 ms ±   1.8 ms    [User: 50.3 ms, System: 12.7 ms]
      Range (min … max):    61.7 ms …  69.7 ms    44 runs

NOTE: This was run multiple times and the results where more or less the
same.